Use floats for light level setting
Nobrainer in hindsight

+++ sim/skylightSim/src/Puddle.cpp
@@ -118,7 +118,7 @@ void Puddle::update_buffers()
 	}
 }
 
-void Puddle::set_light_level(uint8_t level)
+void Puddle::set_light_level(float level)
 {
 	desired_light_level = level;
 }
